+++ src/components/CropRotation/CropRotationYield.tsx
@@ -80,15 +80,15 @@ interface CropFertilizerInfoTable {
 const CropRotationYield: React.FC<{ cropRotationYieldData: CropRotationYieldData[] | null }> = ({
     cropRotationYieldData
 }): JSX.Element => {
-    const { selectedField } = React.useContext(DataStateContext);
+    const { selectedResearch } = React.useContext(DataStateContext);
 
     const [selectedYear, setSelectedYear] = React.useState<string>('');
 
-    const [years, yearsLoading, yearsError] = useAvailableYears(selectedField?.id);
+    const [years, yearsLoading, yearsError] = useAvailableYears(selectedResearch?.id);
 
     const [nitrateConcentrationData, nitrateConcentrationDataLoading, nitrateConcentrationDataError] =
-        useNitrateConcentrationData(selectedYear, selectedField?.id);
-    const [weatherData, weatherDataLoading, weatherDataLoadError] = useWeatherData(selectedYear, selectedField?.id);
+        useNitrateConcentrationData(selectedYear, selectedResearch?.id);
+    const [weatherData, weatherDataLoading, weatherDataLoadError] = useWeatherData(selectedYear, selectedResearch?.id);
 
     const months = [
         'January',